Addressing comments for minor structural changes
