+++ model_wrappers/oncho/testRun.py
@@ -1,15 +1,17 @@
-from post_processing_code.general_functions.aggregation import (
+from post_processing.aggregation import (
     iu_lvl_aggregate,
     country_lvl_aggregate,
     africa_lvl_aggregate,
 )
-from post_processing_code.general_functions.general_post_processing import (
+from post_processing.single_file_post_processing import (
     process_single_file,
-    post_process_file_generator,
-    custom_progress_bar_update,
     measure_summary_float,
 )
-import post_processing_code.constants as constants
+from post_processing.file_util import (
+    post_process_file_generator,
+    custom_progress_bar_update,
+)
+import model_wrappers.constants as constants
 from tqdm import tqdm
 import pandas as pd
 
@@ -19,18 +21,19 @@ file_iter = post_process_file_generator(
     file_directory=oncho_dir, end_of_file="-raw_all_age_data.csv"
 )
 with tqdm(total=1, desc="Post-processing Scenarios") as pbar:
-    for scenario_index, total_scenarios, scenario, country, iu, file_path in file_iter:
+    for file_info in file_iter:
         process_single_file(
-            raw_model_outputs=pd.read_csv(file_path),
-            scenario=scenario,
-            iuName=iu,
+            raw_model_outputs=pd.read_csv(file_info.file_path),
+            scenario=file_info.scenario,
+            iuName=file_info.iu,
             prevalence_marker_name="prevalence",
             post_processing_start_time=1970,
             measure_summary_map={"prevalence": measure_summary_float},
         ).to_csv(
-            "post-processed-outputs/oncho/" + scenario + "_" + iu + "post_processed.csv"
+            "post-processed-outputs/oncho/" + file_info.scenario + "_" +
+            file_info.iu + "post_processed.csv"
         )
-        custom_progress_bar_update(pbar, scenario_index, total_scenarios)
+        custom_progress_bar_update(pbar, file_info.scenario_index, file_info.total_scenarios)
 
 aggregated_df = iu_lvl_aggregate(
     "post-processed-outputs/oncho",
